invalid slice fix

+++ pyqtgraph/graphicsItems/ImageItem.py
@@ -347,8 +347,8 @@ class ImageItem(GraphicsObject):
         if self.image is None:
             return None,None
         if step == 'auto':
-            step = (np.ceil(self.image.shape[0] / targetImageSize),
-                    np.ceil(self.image.shape[1] / targetImageSize))
+            step = (int(np.ceil(self.image.shape[0] / targetImageSize)),
+                    int(np.ceil(self.image.shape[1] / targetImageSize)))
         if np.isscalar(step):
             step = (step, step)
         stepData = self.image[::step[0], ::step[1]]
